refactor: move faq to corresponding pages
