+++ src/api/team.ts
@@ -21,7 +21,7 @@ export function getTeam(): TeamContent {
     card: {
       title: "Styre og team",
       text1:
-        "Vektorprogrammet er en stor organisasjon med assistenter i 4 norske byer. Vi trenger derfor mange frivillige bak kulissene som kan få hjulene til å gå rundt. Uten Vektorprogrammets 15 team hadde dette aldri gått an!",
+        "Vektorprogrammet er en stor organisasjon med assistenter i 4 norske byer. Vi trenger derfor mange frivillige bak kulissene som kan få hjulene til å gå rundt. Uten Vektorprogrammets 15 team hadde dette aldri gått an! ",
       text2: "Kunne du tenkt deg et team-verv hos oss?",
       text3: "Les mer om de ulike teamene nedenfor!",
       image: {
